fix(auth): safeguard the signer in case user reloads page
We have here a dilemma for nsec, if we want to support it.
Our security measure was to save it in the memory, but user can
always reload the page and it would be lost. They would need then
to always insert the nsec, what is not practical.
That means, nsec has to be stored somewhere. This code does it and
applies some measures like deleting it afterwards.